+++ src/components/ProfileTemplate/ProfileTemplate.tsx
@@ -56,7 +56,7 @@ export function ProfileTemplate({userId, isMyProfile}: Props) {
   }
 
   return (
-    <Screen canGoBack={!isMyProfile} flex={1} style={$screen}>
+    <Screen flex={1} style={$screen}>
       <InfinityScrollList
         queryKey={[QueryKeys.PostList, userId]}
         getList={getPostList}
