+++ src/components/Screen/components/ScreenHeader.tsx
@@ -1,8 +1,6 @@
 import React from 'react';
 
-import {useNavigation} from '@react-navigation/native';
-
-import {Box, BoxProps, Icon, Text, TouchableOpacityBox} from '@components';
+import {BackButton, Box, BoxProps, Text} from '@components';
 
 import {ScreenProps} from '../Screen';
 
@@ -15,8 +13,6 @@ export function ScreenHeader({
   HeaderComponent,
   ...boxProps
 }: Props) {
-  const navigation = useNavigation();
-
   if (!title && !canGoBack && !HeaderComponent) {
     return null;
   }
@@ -30,21 +26,7 @@ export function ScreenHeader({
       alignItems="center"
       justifyContent="space-between"
       {...boxProps}>
-      {canGoBack && (
-        <TouchableOpacityBox
-          testID="screen-back-button"
-          flexDirection="row"
-          alignItems="center"
-          mr={showBackLabel ? 's10' : undefined}
-          onPress={navigation.goBack}>
-          <Icon size={ICON_SIZE} name="arrowLeft" color="primary" />
-          {showBackLabel && (
-            <Text preset="paragraphMedium" semiBold ml="s8">
-              Voltar
-            </Text>
-          )}
-        </TouchableOpacityBox>
-      )}
+      {canGoBack && <BackButton showBackLabel={showBackLabel} />}
       {HeaderComponent}
       {title && <Text preset="headingSmall">{title}</Text>}
       {title && <Box backgroundColor="carrotSecondary" width={ICON_SIZE} />}
