+++ src/components/index.ts
@@ -19,3 +19,4 @@ export * from './RadioButton';
 export * from './ProgressIndicator/ProgressIndicator';
 export * from './InfinityScrollList/InfinityScrollList';
 export * from './ProfileTemplate/ProfileTemplate';
+export * from './BackButton/BackButton';
